Move main component out to root
